+++ .github/workflows/build.yml
@@ -27,7 +27,7 @@ jobs:
 
     strategy:
       matrix:
-        node-version: [20.19.5]
+        node-version: [22.21.1]
         # See supported Node.js release schedule at https://nodejs.org/en/about/releases/
 
     steps:
