+++ .github/workflows/publish-oidc.yml
@@ -1,0 +1,54 @@
+# This workflow will run tests using node and then publish a package to npm when a release is created
+# For more information see: https://docs.github.com/en/actions/publishing-packages/publishing-nodejs-packages
+
+name: Node.js Publish with OIDC
+
+on:
+  workflow_dispatch:
+  # release:
+  #   types: [created]
+
+permissions:
+  id-token: write  # Required for OIDC
+  contents: read
+
+# This allows a subsequently queued workflow run to interrupt previous runs
+concurrency:
+  group: '${{ github.workflow }} @ ${{ github.event.pull_request.head.label || github.head_ref || github.ref }}'
+  cancel-in-progress: true
+
+jobs:
+  # build:
+  #   uses: ./.github/workflows/build.yml  
+
+  publish-npm:
+    # needs: build
+    runs-on: ubuntu-22.04
+    strategy:
+      matrix:
+        node-version: [22.21.1]
+        # See supported Node.js release schedule at https://nodejs.org/en/about/releases/
+    steps:
+      - name: Check out repo
+        uses: actions/checkout@v4
+
+      - name: Setup Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: ${{ matrix.node-version }}
+          registry-url: 'https://registry.npmjs.org'
+          cache: npm
+
+      # Ensure npm 11.5.1 or later is installed
+
+      - name: Update npm
+        run: npm install -g npm@latest
+
+      - name: Install dependencies
+        run: npm ci
+
+      - name: Run build only library
+        run: npm run build:prod --if-present
+    
+      - name: Publish package
+        run: npm publish ./dist/ngx-virtual-scroller
