+++ .github/workflows/publish-token.yml
@@ -1,7 +1,7 @@
 # This workflow will run tests using node and then publish a package to npm when a release is created
 # For more information see: https://docs.github.com/en/actions/publishing-packages/publishing-nodejs-packages
 
-name: Node.js Publish
+name: Node.js Publish with Granular Token
 
 on:
   workflow_dispatch:
@@ -19,10 +19,10 @@ jobs:
 
   publish-npm:
     # needs: build
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-22.04
     strategy:
       matrix:
-        node-version: [20.19.5]
+        node-version: [22.21.1]
         # See supported Node.js release schedule at https://nodejs.org/en/about/releases/
     steps:
       - name: Check out repo
@@ -38,7 +38,7 @@ jobs:
       - name: Install dependencies
         run: npm ci
 
-      - name: Run build library
+      - name: Run build only library
         run: npm run build:prod --if-present
     
       - name: Publish package
