+++ angular.json
@@ -10,7 +10,7 @@
       "prefix": "lib",
       "architect": {
         "build": {
-          "builder": "@angular-devkit/build-angular:ng-packagr",
+          "builder": "@angular/build:ng-packagr",
           "options": {
             "project": "projects/ngx-virtual-scroller/ng-package.json"
           },
@@ -43,11 +43,13 @@
       "prefix": "app",
       "architect": {
         "build": {
-          "builder": "@angular-devkit/build-angular:browser",
+          "builder": "@angular/build:application",
           "options": {
-            "outputPath": "dist/demo",
+            "outputPath": {
+              "base": "dist/demo",
+              "browser": ""
+            },
             "index": "projects/demo/src/index.html",
-            "main": "projects/demo/src/main.ts",
             "polyfills": [
               "zone.js"
             ],
@@ -65,7 +67,8 @@
             "allowedCommonJsDependencies": [
               "@tweenjs/tween.js",
               "chance"
-            ]
+            ],
+            "browser": "projects/demo/src/main.ts"
           },
           "configurations": {
             "production": {
@@ -80,7 +83,7 @@
           "defaultConfiguration": "production"
         },
         "serve": {
-          "builder": "@angular-devkit/build-angular:dev-server",
+          "builder": "@angular/build:dev-server",
           "configurations": {
             "production": {
               "buildTarget": "demo:build:production"
