+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ngx-virtual-scroller",
-  "version": "4.0.9",
+  "version": "4.0.10",
   "scripts": {
     "ng": "ng",
     "build": "npm-run-all -s clean:build build:prod",
@@ -19,32 +19,35 @@
     "lint:lib": "ng lint @talentia/ngx-virtual-scroller",
     "lint:app": "ng lint demo",
     "lint:lib-fix": "ng lint @talentia/ngx-virtual-scroller --fix",
-    "lint:app-fix": "ng lint demo --fix"
+    "lint:app-fix": "ng lint demo --fix",
+    "madge:app": "madge --circular --ts-config ./projects/demo/tsconfig.app.json --extensions ts,html ./projects/demo/src/",
+    "madge:lib": "madge --circular --ts-config ./projects/ngx-virtual-scroller/tsconfig.lib.prod.json --extensions ts ./projects/ngx-virtual-scroller/",
+    "update-browserslist": "npx update-browserslist-db@latest"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "18.2.14",
-    "@angular/common": "18.2.14",
-    "@angular/compiler": "18.2.14",
-    "@angular/core": "18.2.14",
-    "@angular/forms": "18.2.14",
-    "@angular/platform-browser": "18.2.14",
-    "@angular/platform-browser-dynamic": "18.2.14",
-    "@angular/router": "18.2.14",
+    "@angular/animations": "19.2.15",
+    "@angular/common": "19.2.15",
+    "@angular/compiler": "19.2.15",
+    "@angular/core": "19.2.15",
+    "@angular/forms": "19.2.15",
+    "@angular/platform-browser": "19.2.15",
+    "@angular/platform-browser-dynamic": "19.2.15",
+    "@angular/router": "19.2.15",
     "@tweenjs/tween.js": "17.4.0",
     "chance": "1.1.4",
     "rxjs": "7.8.1",
     "tslib": "2.6.2",
-    "zone.js": "0.14.10"
+    "zone.js": "0.15.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "18.2.21",
-    "@angular/cli": "18.2.21",
-    "@angular/compiler-cli": "18.2.14",
+    "@angular/build": "^19.2.19",
+    "@angular/cli": "19.2.19",
+    "@angular/compiler-cli": "19.2.15",
     "@types/sortablejs": "1.15.0",
-    "angular-eslint": "18.4.3",
+    "angular-eslint": "19.8.1",
     "eslint": "9.15.0",
-    "ng-packagr": "18.2.1",
+    "ng-packagr": "19.2.2",
     "npm-run-all": "4.1.5",
     "rimraf": "6.0.1",
     "typescript": "5.5.2",
