+++ projects/demo/src/app/lists/horizontal-list.component.ts
@@ -1,11 +1,14 @@
 import { Component,Input } from '@angular/core';
+import { VirtualScrollerModule } from '@talentia/ngx-virtual-scroller';
+import { NgFor, NgIf } from '@angular/common';
 import { BaseListComponent } from './base-list.component';
+import { ListItemComponent } from './list-item.component';
 
 @Component({
-  standalone: false,
-  selector: 'horizontal-list',
-  templateUrl: 'horizontal-list.component.html',
-  styleUrls: ['./horizontal-list.scss']
+    selector: 'horizontal-list',
+    templateUrl: 'horizontal-list.component.html',
+    styleUrls: ['./horizontal-list.scss'],
+    imports: [NgIf, NgFor, VirtualScrollerModule, ListItemComponent]
 })
 export class HorizontalListComponent extends BaseListComponent {
   @Input() rtl: boolean;
