+++ projects/demo/src/app/lists/multi-col-list.component.ts
@@ -1,11 +1,15 @@
 import { Component } from '@angular/core';
+import { NgFor, NgIf } from '@angular/common';
+import { VirtualScrollerModule } from '@talentia/ngx-virtual-scroller';
+import { ListItemComponent } from './list-item.component';
 import { BaseListComponent } from './base-list.component';
 
+
 @Component({
-  standalone: false,
-  selector: 'multi-col-list',
-  templateUrl: 'multi-col-list.component.html',
-  styleUrls: ['./multi-col-list.scss']
+    selector: 'multi-col-list',
+    templateUrl: 'multi-col-list.component.html',
+    styleUrls: ['./multi-col-list.scss'],
+    imports: [NgIf, NgFor, VirtualScrollerModule, ListItemComponent]
 })
 export class MultiColListComponent extends BaseListComponent {
 }
