+++ projects/demo/src/app/lists/table-list.component.ts
@@ -1,11 +1,13 @@
 import { Component } from '@angular/core';
+import { NgFor } from '@angular/common';
+import { VirtualScrollerModule } from '@talentia/ngx-virtual-scroller';
 import { BaseListComponent } from './base-list.component';
 
 @Component({
-  standalone: false,
-  selector: 'table-list',
-  templateUrl: 'table-list.component.html',
-  styleUrls: ['./table-list.scss']
+    selector: 'table-list',
+    templateUrl: 'table-list.component.html',
+    styleUrls: ['./table-list.scss'],
+    imports: [NgFor, VirtualScrollerModule]
 })
 export class TableListComponent extends BaseListComponent {
 }
