+++ projects/demo/src/app/parent/parent-scroll.component.ts
@@ -1,10 +1,13 @@
 import { Component } from '@angular/core';
+import { NgIf } from '@angular/common';
+import { VirtualScrollerModule } from '@talentia/ngx-virtual-scroller';
 import { BaseListComponent } from '../lists/base-list.component';
+import { ListItemComponent } from '../lists/list-item.component';
 
 @Component({
-  standalone: false,
-  selector: 'parent-scroll',
-  templateUrl: 'parent-scroll.component.html'
+    selector: 'parent-scroll',
+    templateUrl: 'parent-scroll.component.html',
+    imports: [NgIf,VirtualScrollerModule, ListItemComponent]
 })
 export class ParentScrollComponent extends BaseListComponent {
   constructor() {
