+++ projects/demo/src/app/samples/samples.component.ts
@@ -1,12 +1,17 @@
 import { Component } from '@angular/core';
 import { ListItem } from '../lists/list-item.component';
 import { BaseListComponent } from '../lists/base-list.component';
+import { VerticalListComponent } from '../lists/vertical-list.component';
+import { HorizontalListComponent } from '../lists/horizontal-list.component';
+import { MultiColListComponent } from '../lists/multi-col-list.component';
+import { TableListComponent } from '../lists/table-list.component';
+import { ListWithApiComponent } from '../lists/list-with-api.component';
 
 @Component({
-  standalone: false,
-  selector: 'samples',
-  templateUrl: 'samples.component.html',
-  styleUrls: ['./samples.component.scss']
+    selector: 'samples',
+    templateUrl: 'samples.component.html',
+    styleUrls: ['./samples.component.scss'],
+    imports: [VerticalListComponent, HorizontalListComponent, MultiColListComponent, TableListComponent, ListWithApiComponent]
 })
 export class SamplesComponent {
   public items: ListItem[] = [];
