+++ projects/ngx-virtual-scroller/eslint.config.js
@@ -7,6 +7,7 @@ module.exports = tseslint.config(
   {
     files: ["**/*.ts"],
     rules: {
+      "@angular-eslint/prefer-standalone": "off",
       "@typescript-eslint/no-explicit-any": "off",
       "@typescript-eslint/no-this-alias": "off",
       "@typescript-eslint/no-unsafe-function-type": "off",
