+++ projects/ngx-virtual-scroller/package.json
@@ -1,12 +1,12 @@
 {
   "name": "@talentia/ngx-virtual-scroller",
-  "version": "4.0.9",
+  "version": "4.0.10",
   "description": "Angular module for virtual infinite list. Supports horizontal/vertical, variable heights & multi-column",
   "license": "MIT",
   "licenseFilename": "LICENSE",
   "peerDependencies": {
-    "@angular/common": "^18.0.0",
-    "@angular/core": "^18.0.0",
+    "@angular/common": "^19.0.0",
+    "@angular/core": "^19.0.0",
     "@tweenjs/tween.js": "~17.4.0" 
   },
   "author": {
